Change tick_t to int64_t to prevent potential overflows

+++ include/lmms_basics.h
@@ -37,7 +37,7 @@
 
 
 typedef int32_t bar_t;
-typedef int32_t tick_t;
+typedef int64_t tick_t;
 typedef uint8_t volume_t;
 typedef int8_t panning_t;
